Stock_Hub_new/New/requirements.txt
absl-py==2.1.0
altair==5.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlover==0.3.0
comm==0.2.2
contourpy==1.2.1
cufflinks==0.17.3
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.4
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.65.5
h5py==3.11.0
html5lib==1.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
multitasking==0.0.11
namex==0.0.8
narwhals==1.4.2
nest-asyncio==1.6.0
newsapi-python==0.2.7
nltk==3.9
numpy==1.26.4
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pandas-datareader==0.10.0
pandas-ta==0.3.14b0
parso==0.8.4
peewee==3.17.5
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.1.1
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.37.1
ta==0.11.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
yahoofinance==0.0.2
yfinance==0.2.41
